requirements.txt
pandas
numpy
matplotlib
seaborn
langchain==0.3.11
langchain-openai==0.2.12
langchain-community==0.3.11
huggingface_hub==0.26.5
transformers==4.46.3
python-dotenv==1.0.1
streamlit==1.39.0
